xkyy/requirement.txt
aiohttp==3.5.4
aiomysql==0.0.20
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appdirs==1.4.3
Appium-Python-Client==0.40
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
autopep8==1.4.3
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
baidu-aip==2.2.12.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
CacheControl==0.12.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
codes==0.1.0
colorama==0.3.9
comtypes==1.1.2
conda==4.6.8
conda-build==3.0.27
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.0.3
cssselect==1.0.3
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.0.11
defusedxml==0.5.0
demjson==2.2.4
distlib==0.2.5
distributed==1.19.1
Django==2.2.1
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-formtools==2.1
django-js-asset==1.2.2
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
docopt==0.6.2
docutils==0.14
dukpy==0.2.2
echarts-countries-pypkg==0.1.6
entrypoints==0.2.3
et-xmlfile==1.0.1
facebook-wda==0.2.2
fake-useragent==0.1.11
fastcache==1.0.2
filelock==2.0.12
fire==0.1.3
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Login==0.4.1
future==0.17.1
gevent==1.2.2
glob2==0.5
greenlet==0.4.12
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
humanize==0.5.1
hyperlink==18.0.0
idna==2.6
idna-ssl==1.1.0
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
javascripthon==0.10
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
Js2Py==0.60
jsmin==2.2.2
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyter-echarts-pypkg==0.1.2
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
lml==0.0.2
locket==0.2.0
lockfile==0.12.2
logzero==1.5.0
lxml==4.1.0
macropy3==1.1.0b2
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.16
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multidict==4.5.2
multipledispatch==0.4.9
mysqlclient==1.4.2.post1
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
parse==1.11.1
parsel==1.5.1
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pdfkit==0.6.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
ply==3.10
prettytable==0.7.2
progress==1.3
prompt-toolkit==1.0.15
psutil==5.4.0
py==1.4.34
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
pyee==5.0.0
PyExecJS==1.5.1
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
pyHook==1.5.1
pyjsparser==2.5.2
pylint==1.7.4
pymongo==3.7.2
PyMySQL==0.9.2
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.6.7
pysolr==3.8.1
pyspider==0.3.10
pytest==3.2.1
python-dateutil==2.6.1
pytz==2017.3
PyUserInput==0.1.11
PyWavelets==0.5.2
pywifi==1.1.12
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
requests-html==0.10.0
retry==0.9.2
rope==0.10.5
ruamel-yaml==0.11.14
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
Scrapy==1.6.0
scrapy-redis==0.6.8
seaborn==0.8
selenium==3.141.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
tqdm==4.11.2
traitlets==4.3.2
Twisted==18.9.0
typing==3.6.2
typing-extensions==3.7.2
tzlocal==1.5.1
u-msgpack-python==2.5.1
uiautomator2==0.2.3.dev2
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.4.3
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
weditor==0.1.2.dev1
Werkzeug==0.12.2
whichcraft==0.5.2
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
WsgiDAV==2.4.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
yarl==1.3.0
zict==0.1.3
zope.interface==4.6.0
